Moved MAX_CACHE_SIZE to priv struct.
The MAX_CACHE_SIZE is no longer global but is now private data.
It is a guint64 because it is compared against the cache_size
which is computed by a Glib function which returns a guint64.
[endlessm/eos-sdk#1547]

+++ daemon/emer-persistent-cache.h
@@ -133,7 +133,7 @@ gboolean             emer_persistent_cache_store_metrics       (EmerPersistentCa
 EmerPersistentCache *emer_persistent_cache_new_full            (GCancellable             *cancellable,
                                                                 GError                  **error,
                                                                 gchar                    *custom_directory,
-                                                                gint                      custom_cache_size,
+                                                                guint64                   custom_cache_size,
                                                                 EmerBootIdProvider       *boot_id_provider,
                                                                 EmerCacheVersionProvider *version_provider,
                                                                 guint                     boot_offset_update_interval);
